Create Array1.java
GFG day 1 challenge: find the second largest number in array.
[Naive Approach] Using Sorting - O(n*logn) Time and O(1) Space
The idea is to sort the array in non-decreasing order. Now, we know that the largest element will be at index n - 1. So, starting from index (n - 2), traverse the remaining array in reverse order. As soon as we encounter an element which is not equal to the largest element, return it as the second largest element in the array. If all the elements are equal to the largest element, return -1.